Fix: Support Info module provides no GitGutter version.
The issue was introduced with commit 3ba34d4.

The VERSION file no longer existed and could no longer be read by the support module.

This commit adds the VERSION which is generated during release to the root of the package.
The 'messages' folder must only contain the install.txt and the release messages. Otherwise the current release process would fail.