Enhancement: Improve diff popup visibility on light color schemes
The `background-color` html attribute seems useless and is removed.
Therefore a new color `--diffpopup-tb-bg` for the toolbar background
is added which adapts to the color scheme.

Default hyperlinks are highlighted light blue and therefore barely
visible on light color schemes. With this commit the text foreground
color is used instead. Disabled links (toolbar text) is blended
between fore- and background color.