Fix: Forwared arguments from commands to git_gutter
The GitGutterBaseCommand acts as proxy to forward all commands to the main
GitGutterCommand. This commit enables argument forwarding for this proxy.

Prior to this commit all commands using arguments, lost them at this point.
This issue effected the following commands:

- revert_change(line=None, point=None)
- copy_from_commit(line=None, point=None)

As both functions use the cursor position if none of the arguments is
available, it was not detected earlier.