Enhancement: Add more status bar variables (Issue #457)
This commit replaces the `get_current_branch()` by `git_branch_status()`.

The new method calls `git status -b -s -u` to get general status and
statistics of the current branch and working tree, which are passed to
as variables to jinja2 to be used for customized status bar messages.

The new variables are:

  {{remote}}         -- tracked remote of current branch you are working on or `None`
  {{ahead}}          -- number of commits the local branch is ahead of remote
  {{behind}}         -- number of commits the local branch is behind remote
  {{added_files}}    -- number of untracked files added to working tree
  {{deleted_files}}  -- number of files deleted from working tree
  {{modified_files}} -- number of modified files in the working tree
  {{staged_files}}   -- number of files in the staging area

Note: The variables are not used in the default templates.

The method is used for "Compare against origin" as well and is therefore
able to output an error message if the current branch does not have an
tracking remote attached.