Fix: mangled {{branch}} text when `git status` output is colored (Issue #474)
* add "-c color.status=never" option to `git status`, definitively disabling color output

.### Discussion

{{branch}} is mis-parsed if `git status` output is colored (eg, when `git` is configured
with "config.status=always"). Using the "-c color.status=never" option overrides any user
configuration, displaying `git status` as uncolored in all cases and fixing the issue.

Notably, option ordering is important in this instance. The "-c ..." option is a `git`
base option, not a `git status` subcommand option. So, as done here, it must be inserted
before the "status" subcommand.