Enhancement: Place all temporary files into a dedicated sub directory.
Fixes issue #483

Introduces the `TempFile` class to handle access to temporary files and to ensure file deletion.

A look into python's tempfile module revealed a possible issue with temporary files not being removed on unix/linux/mac if `os.unlink()` is called within the `__del__()` method of an object. The `os` module might no longer exist at the point of time the `__del__()` is called especially upon application shutdown.

To ensure the files to be deleted, it is required to cache the `os.unlink()` function within the `TempFile` class and call the cached one within `__del__()`.