Internal: Use XDG_RUNTIME_DIR for temporary files if available
The XDG_RUNTIME_DIR on unix based systems is meant to store user specific runtime files. The folder is created at logon and cleared, if a user logs off. This way left behind files are removed for sure.

If the environment variable does not exist, GitGutter falls back using the normal TEMP directory. If available the UID is added, which helps on Linux to create a per user temporary directory.

On Windows TEMP points to a user specific folder by default. So adding a UID to the path may not be required.