Enhancement: Delete old temporary files upon startup
In some cases temporary files don't seem to be deleted, if ST is closed. To ensure old files not to flood the temporary directory, all files which haven't been accessed for 2 days are removed upon startup.

Remark: Operating systems which support the XDG_RUNTIME_DIR don't need it as this directory is cleared out each time a user loggs off, but it is handled as ordinary tempfile to ensure files to be removed if a user doesn't log off for a long time for some reason.