Fix: ST2 compatibility issue in TempFile class
Python 2.6 does not support the `FileExistsError` exception being used to catch
the already existing temporary directory, which is tried to be created each time
a temporary `TempFile` instance is initialized.

Therefore the `OSError` is catched and the return value of `mkdir` is evaluated.

The `_file` attribute is initialized before `mkdir`, because of an
`AttributeError` being raised while deleting the `TempFile` instance.