Fix: Check git version after settings change.
This commit fixes an issue which could cause GitGutter to assume the
wrong version of the git binary after changing the "git_binary" setting.

The git binary is therefore cached by the GitGutterHandler object of a
view and used to trigger a version check if the "git_binary" has changed.

All subsequent methods now use the cached and validated git binary
to ensure consistency.