Enhancement: Directly write git cache file to disk
This commit directly writes the output of the `git cat-file` to the
temporary cache without tunneling it through stdout. Therefore less
RAM is needed for this operation and GitGutter no longer needs to
struggle with the whole file content.

The subprocess is created directly and runs in the background while
being polled repeatedly until finished. This allows the plugin_host
to do other stuff while `git cat-file` is busy even though no thread
is spawned to wait for the process to finish. This should save some
resources and the time needed to spawn the thread.

This change is furthermore needed in preparation for WSL support as
the bash.exe / wsl.exe often truncates the output of `git cat-file`
if read via stdout.

This commit adds an `PromiseError` object to be used to resolve a
promise in cases of caught exceptions or errors in order to let
the following function handle it gracefully. This concept might be
used in other situations in the future, too.