Internal: Directly start asynchronous processes
This commit uses the strategy to directly spawn a subprocess and use
the poll mechanism to check for its finish from the previous commit.

This way the time the plugin_host or the main thread on ST is blocked
for waiting for the process to end is limited.

As output is read from stdout it needs to be read during each poll as
otherwise the process won't exit if the reading buffer is full.