Fix: Skip evaluation if rebase is active (Issue #508)
If GitGutter runs `git status` while a rebase is on the fly the index
might become locked which causes the rebase operation to fail.

Furthermore the files of a repository may change very often during
rebasing which doesn't need to be displayed anyway.

To avoid displaying wrong information or run 'actions' while rebasing
is active, GitGutter is disabled for the meanwhile.

It will be enabled as soon as the rebase was finished and the view
is activated or modified again.