Internal: Add Template Cache Manager
This commit introduces the `templates` module which acts as intermediate layer between GitGutter and jinja2 library.

If jinja2 is available the module creates a weakref dictionary to hold all created templates. The function `create` tries to return an already existing template instance first. If none was found in the cache a new one is created and cached for later use.

Jinja2 can do such caching based on template names only. GitGutter uses the source from settings as `key` to distinguish the templates as names are equal for each view.