Feature: Display git blame message of active line (#487)
GitGutter learned to run git blame for the active line and display the results in the status bar and as phantom text at the end of the line.

Both outputs can be enabled/disabled and customized via jinja2 templates.

The color of the line annotation is to be defined in the color scheme by adding a rule for the scope `comment.line.annotation.git_gutter`.