Internal: Run git blame only if used by a template
This commit adds an evaluation algorithm for the user templates to check whether specific variables are used or not. This is used to run git only if required. If `show_line_annotation` is set False and the status bar template does not contain any `line_...` variables `git blame` is skipped.