Fix: Invalid key binding templates (#530)
If `Preferences: GitGutter Key Bindings` is called in a vanilla ST setup, a new `Package/User/Default (...).sublime-keymap` file is created containing an invalid JSON template. Hence loading keymaps fails with an parse error.

This commit fixes the template to add an list rather than an object to the new file. As the created file is a user's default file the name `GitGutter` is removed as well.