Enhancement: Don't add gutter icons, if ST's mini diff is enabled
ST 3188+ introduces a mini diff in the gutter, which is a more efficient alternative to provide info about added/deleted/modified lines than GitGutter can ever do.

This commit queries the new `mini_diff` setting to decide, whether to show gutter icons or not.

Anything else keeps as is so far, as Sublime Text does not yet show mini diff infos in the minimap. So GitGutter still needs to add the regions as is to not break the minimap feature.