Enhancement: Read branch statistics on demand
Workaround for Issue #546

This commit tries to avoid calling `git status -b -s` if the status bar
text template does not indicate to need those information as this
command is quite expensive especially when working with repos on a
network share.

In order to avoid calling `git status` the following variables must not be present in the
  `status_bar_text` (GitGutter.sublime-settings) or
  `git_gutter_status_bar_text` (Preferences.sublime-settings):

- {{remote}}
- {{ahead}}
- {{behind}}
- {{added_files}}
- {{deleted_files}}
- {{modified_files}}
- {{staged_files}}

If {{branch}} is present in the template, GitGutter calls
`git rev-parse --abbrev-ref HEAD` to read the branch name.