Fix: A saved as.. file is compared against the original one (#549)
Fixes #549

This commit fixes an issue, which caused a renamed or saved as... file to be compared against the original object in the commit.

Even though the saved file might no longer be tracked, GitGutter still
displayed it as modified or committed and displayed modifications in
the gutter as if the file was not renamed.