Fix: Empty compare target
Fixes #560

Setting compare target to `origin` may cause a `None` target to be
applied. GitGutter no longer knows that to compare against and raises
exceptions.

This commit fixes `on_branch_name()` to set a new compare target if the
given value is empty and ensures the handler to use sane fallback values
in case compare target values are empty.